+++ docs-website/src/pages/_document.jsx
@@ -1,0 +1,149 @@
+import { Head, Html, Main, NextScript } from 'next/document'
+
+const themeScript = `
+  let mediaQuery = window.matchMedia('(prefers-color-scheme: dark)')
+
+  function updateTheme(savedTheme) {
+    let theme = 'system'
+    try {
+      if (!savedTheme) {
+        savedTheme = window.localStorage.theme
+      }
+      if (savedTheme === 'dark') {
+        theme = 'dark'
+        document.documentElement.classList.add('dark')
+      } else if (savedTheme === 'light') {
+        theme = 'light'
+        document.documentElement.classList.remove('dark')
+      } else if (mediaQuery.matches) {
+        document.documentElement.classList.add('dark')
+      } else {
+        document.documentElement.classList.remove('dark')
+      }
+    } catch {
+      theme = 'light'
+      document.documentElement.classList.remove('dark')
+    }
+    return theme
+  }
+
+  function updateThemeWithoutTransitions(savedTheme) {
+    updateTheme(savedTheme)
+    document.documentElement.classList.add('[&_*]:!transition-none')
+    window.setTimeout(() => {
+      document.documentElement.classList.remove('[&_*]:!transition-none')
+    }, 0)
+  }
+
+  document.documentElement.setAttribute('data-theme', updateTheme())
+
+  new MutationObserver(([{ oldValue }]) => {
+    let newValue = document.documentElement.getAttribute('data-theme')
+    if (newValue !== oldValue) {
+      try {
+        window.localStorage.setItem('theme', newValue)
+      } catch {}
+      updateThemeWithoutTransitions(newValue)
+    }
+  }).observe(document.documentElement, { attributeFilter: ['data-theme'], attributeOldValue: true })
+
+  mediaQuery.addEventListener('change', updateThemeWithoutTransitions)
+  window.addEventListener('storage', updateThemeWithoutTransitions)
+`
+
+export default function Document() {
+	return (
+		<Html className="antialiased [font-feature-settings:'ss01']" lang="en">
+			<Head>
+				<script dangerouslySetInnerHTML={{ __html: themeScript }} />
+				<link
+					rel="apple-touch-icon"
+					sizes="57x57"
+					href="/favicon/apple-icon-57x57.png"
+				/>
+				<link
+					rel="apple-touch-icon"
+					sizes="60x60"
+					href="/favicon/apple-icon-60x60.png"
+				/>
+				<link
+					rel="apple-touch-icon"
+					sizes="72x72"
+					href="/favicon/apple-icon-72x72.png"
+				/>
+				<link
+					rel="apple-touch-icon"
+					sizes="76x76"
+					href="/favicon/apple-icon-76x76.png"
+				/>
+				<link
+					rel="apple-touch-icon"
+					sizes="114x114"
+					href="/favicon/apple-icon-114x114.png"
+				/>
+				<link
+					rel="apple-touch-icon"
+					sizes="120x120"
+					href="/favicon/apple-icon-120x120.png"
+				/>
+				<link
+					rel="apple-touch-icon"
+					sizes="144x144"
+					href="/favicon/apple-icon-144x144.png"
+				/>
+				<link
+					rel="apple-touch-icon"
+					sizes="152x152"
+					href="/favicon/apple-icon-152x152.png"
+				/>
+				<link
+					rel="apple-touch-icon"
+					sizes="180x180"
+					href="/favicon/apple-icon-180x180.png"
+				/>
+				<link
+					rel="icon"
+					type="image/png"
+					sizes="192x192"
+					href="/favicon/android-icon-192x192.png"
+				/>
+				<link
+					rel="icon"
+					type="image/png"
+					sizes="32x32"
+					href="/favicon/favicon-32x32.png"
+				/>
+				<link
+					rel="icon"
+					type="image/png"
+					sizes="96x96"
+					href="/favicon/favicon-96x96.png"
+				/>
+				<link
+					rel="icon"
+					type="image/png"
+					sizes="16x16"
+					href="/favicon/favicon-16x16.png"
+				/>
+				<link rel="manifest" href="/favicon/manifest.json" />
+				<meta name="msapplication-TileColor" content="#ffffff" />
+				<meta
+					name="msapplication-TileImage"
+					content="/favicon/ms-icon-144x144.png"
+				/>
+				<meta name="theme-color" content="#ffffff" />
+				<link
+					rel="preload"
+					href="/fonts/virgil.woff2"
+					as="font"
+					type="font/woff2"
+					crossOrigin="anonymous"
+				/>
+			</Head>
+			<body className="bg-white dark:bg-gray-950">
+				<Main />
+				<NextScript />
+			</body>
+		</Html>
+	)
+}